# workspace/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NetOps Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /provisioning/preview:
    post:
      operationId: previewProvisioningConfig
      tags: [provisioning-jobs]
      summary: Generate config preview (serviceType workflow or templateId v0.4 catalog)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/ProvisioningPreviewInput"
                - $ref: "#/components/schemas/ProvisioningPreviewRequest"
      responses:
        "200":
          description: Preview result
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ProvisioningPreviewResult"
                  - $ref: "#/components/schemas/ProvisioningPreviewResponse"
        "400":
          description: Invalid input
        "404":
          description: Device not found

  # ── PROVISIONING JOBS ────────────────────────────────────────────────────────
...
  /provisioning-jobs:
    get:
      operationId: listProvisioningJobs
      tags: [provisioning-jobs]
      summary: List provisioning jobs
      parameters:
        - name: deviceId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProvisioningJob"
    post:
      operationId: createProvisioningJob
      tags: [provisioning-jobs]
      summary: Create a provisioning job (L2VPN/L3VPN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProvisioningJobInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvisioningJob"
...
  /provisioning-jobs/{id}:
    get:
      operationId: getProvisioningJob
      tags: [provisioning-jobs]
      summary: Get provisioning job details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvisioningJobDetail"
  /provisioning-jobs/{id}/validate:
    post:
      operationId: validateProvisioningJob
      tags: [provisioning-jobs]
      summary: Validate a provisioning job before execution
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Validation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationResult"
  /provisioning-jobs/{id}/preview:
    post:
      operationId: previewProvisioningJob
      tags: [provisioning-jobs]
      summary: Export Markdown plan for an existing job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job preview with Markdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvisioningJobPreviewResponse"
  /provisioning-jobs/{id}/request-approval:
    post:
      operationId: requestProvisioningJobApproval
      tags: [provisioning-jobs]
      summary: Move validated job to pending_approval
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job pending approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvisioningJobDetail"
        "409":
          description: Invalid status transition
  /provisioning-jobs/{id}/approve:
    post:
      operationId: approveProvisioningJob
      tags: [provisioning-jobs]
      summary: Approve job after pending_approval (apply still blocked by default)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvisioningJobDetail"
        "409":
          description: Invalid status transition
  /provisioning-jobs/{id}/cancel:
    post:
      operationId: cancelProvisioningJob
      tags: [provisioning-jobs]
      summary: Cancel draft/validated/pending job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cancelled job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvisioningJobDetail"
        "409":
          description: Invalid status transition
  /provisioning-jobs/{id}/execute:
    post:
      operationId: executeProvisioningJob
      tags: [provisioning-jobs]
      summary: Execute a provisioning job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvisioningJobDetail"
  /provisioning-jobs/{id}/rollback:
    post:
      operationId: rollbackProvisioningJob
      tags: [provisioning-jobs]
      summary: Rollback a provisioning job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rollback result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProvisioningJobDetail"

  # ── COLLECTED CONFIGS ─────────────────────────────────────────────────────────
...
  /provisioning-jobs/{id}/report:
    post:
      operationId: createProvisioningReport
      tags: [provisioning-jobs]
      summary: Generate a provisioning report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Generated report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
...
components:
  schemas:
...
    ProvisioningPreviewInput:
      type: object
      required: [deviceId, serviceType, parameters]
      properties:
        deviceId:
          type: integer
        serviceType:
          type: string
        parameters:
          type: object
          additionalProperties: true
        maintenanceWindowStart:
          type: ["string", "null"]
        maintenanceWindowEnd:
          type: ["string", "null"]
        rollbackPlan:
          type: ["string", "null"]
    ProvisioningPreviewValidation:
      type: object
      required: [name, passed, message]
      properties:
        name:
          type: string
        passed:
          type: boolean
        message:
          type: string
        severity:
          type: string
    ProvisioningPreviewResult:
      type: object
      required: [deviceId, serviceType, configPreview, rollbackPreview, validations, risks, missingData, applyBlocked]
      properties:
        deviceId:
          type: integer
        serviceType:
          type: string
        configPreview:
          type: string
        rollbackPreview:
          type: string
        validations:
          type: array
          items:
            $ref: "#/components/schemas/ProvisioningPreviewValidation"
        risks:
          type: array
          items:
            type: string
        missingData:
          type: array
          items:
            type: string
        maintenanceWindow:
          type: ["object", "null"]
          properties:
            start:
              type: ["string", "null"]
            end:
              type: ["string", "null"]
        rollbackPlan:
          type: ["string", "null"]
        applyBlocked:
          type: boolean
        applyBlockedReason:
          type: ["string", "null"]
...
    ProvisioningPreviewRequest:
      type: object
      required: [deviceId, templateId]
      properties:
        deviceId:
          type: integer
        templateId:
          type: string
        parameters:
          type: object
          additionalProperties: true
        mode:
          type: string
          default: dry_run
        maintenanceWindowStart:
          type: ["string", "null"]
        maintenanceWindowEnd:
          type: ["string", "null"]
        rollbackPlan:
          type: ["string", "null"]
    ProvisioningPreviewResponse:
      type: object
      required:
        - status
        - deviceId
        - templateId
        - serviceType
        - configPreview
        - rollbackPreview
        - executionPlan
        - validations
        - risks
        - precheckHints
        - postcheckHints
        - missingData
        - blockedReasons
        - applyBlocked
      properties:
        status:
          type: string
          enum: [valid, warning, blocked]
        deviceId:
          type: integer
        templateId:
          type: string
        serviceType:
          type: string
        configPreview:
          type: string
        rollbackPreview:
          type: string
        executionPlan:
          type: array
          items:
            type: string
        validations:
          type: array
          items:
            $ref: "#/components/schemas/ProvisioningValidationItem"
        risks:
          type: array
          items:
            $ref: "#/components/schemas/ProvisioningRisk"
        precheckHints:
          type: array
          items:
            type: string
        postcheckHints:
          type: array
          items:
            type: string
        missingData:
          type: array
          items:
            type: string
        blockedReasons:
          type: array
          items:
            type: string
        applyBlocked:
          type: boolean
        applyBlockedReason:
          type: ["string", "null"]
        maintenanceWindow:
          type: ["object", "null"]
          properties:
            start:
              type: ["string", "null"]
            end:
              type: ["string", "null"]
        rollbackPlan:
          type: ["string", "null"]
...
    ProvisioningJobPreviewResponse:
      type: object
      properties:
        previewMarkdown:
          type: string

    # ── PROVISIONING JOBS ──────────────────────────────────────────────────────
    ProvisioningJob:
      type: object
      required: [id, name, type, status, deviceIds, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          description: "l2vpn_vpws, l2vpn_vpls, l3vpn_vrf, bgp_peer_customer, bgp_peer_provider, l2vpn, l3vpn"
        status:
          type: string
          description: "draft, validated, pending_approval, approved, blocked, cancelled, executing, completed, failed, rolled_back"
        deviceIds:
          type: array
          items:
            type: integer
        templateId:
          type: ["integer", "null"]
        parameters:
          type: ["string", "null"]
          description: "JSON of provisioning parameters"
        validatedAt:
          type: ["string", "null"]
        executedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
    ProvisioningJobInput:
      type: object
      required: [name, type, deviceIds]
      properties:
        name:
          type: string
        type:
          type: string
        deviceIds:
          type: array
          items:
            type: integer
        templateId:
          type: integer
        parameters:
          type: string
    ProvisioningJobDetail:
      type: object
      required: [id, name, type, status, deviceIds, steps, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
        status:
          type: string
        deviceIds:
          type: array
          items:
            type: integer
        templateId:
          type: ["integer", "null"]
        parameters:
          type: ["string", "null"]
        validatedAt:
          type: ["string", "null"]
        executedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/ProvisioningStep"
...
    ValidationResult:
      type: object
      required: [valid, checks]
      properties:
        valid:
          type: boolean
        checks:
          type: array
          items:
            $ref: "#/components/schemas/ValidationCheck"
...